Make sure language locale always changes to the appropriate system language.
